Added ParserEvents. parse() method no more abstract. Other changes.
* ParserEvents can be added. This is the way, by which parser errors is
reported to the end interface. This is done, because, the normal use of
parse() method will be inside a try-catch and on error,
moveToNextBlock() will be executed. This means, it will the Parser
errors will get hidden from the interface, by the classes which
actually class parse().

* ParsingErros enum now moved from KindleParserV1 to AbstractParser.
This was done to accomodate other changes, which implement the parse()
method in the AbstractParser, and only part absolutly involved in the
actual parsing is now abstract. This makes sub-classing easy, as the
calling methods from ParserEvents, throwing erros are handled by the
super class itself.

* Related to the above change, some hooker methods is added and called
in parse() method.

* Some code formatting changes, to make it more readable.