Replacing UseJava with custom script
Previous solution:

With UseJava (which comes with CMake v3.18), the resource files are
added to wrong location. To remedy this, I previously used to add the
resource flies after the initial jar file was created by UseJava. But
the implementation bugs:

> 1. The `add_custom_command` which adds resource files to jar files is
     executed every time. This is slowing down building process.
> 2. At this point, building using cmake is taking same/more time than when
    using build.sh

Both these problems was solved in this commit.

Current Solution:

UseJava is replaced with a custom script which compiles java source files
and creates jar flies with resource and manifest files. The resource files
are also placed in proper place.

With this change, steps are only executed if required not if there is
no change between builds, nothing gets built. It therefore builds much
faster than build.sh

Current Issues:
1. Parallel builds are failing.
   Always use simple `make` or `cmake -j1 --build .` for building.
2. Script needs to be made more tidy and small.