Fix: Package not rebuild when dependency updates
Issue:
When files of a target changes, its target gets rebuild, but not the
targets which depend on it. For example: When a function in AJL is
modified, targets AJL, KPDFSYNC.LIB and KPDFSYNC.GUI should also
rebuild, but only AJL and AJL.jar was rebuild. This is of course
wrong and may cause fault at run time.

Solution:
Adds the JAR file of parent dependencies when building child targets.
Here AJL is the parent dependency for both KPDFSYNC.LIB and KPDFSYNC.GUI
targets, so AJL.JAR file is mention in the DEPENDS of of the custom
command which builds the class files in KPDFSYNC.LIB and KPDFSYNC.GUI
targets.

Also in this commit:
* New DEPENDS argument in `add_jar` (provides fix for the above issue)
* The DEPENDS makes `add_dependencies` command redundant thus removed.
* Other minor changes.