Cross platform version string for each package.
Fixes previous drawbacks of version string creation.

Previous Drawbacks:
> Only works in Linux as of now.
Now works on Windows as well.

> GIT absence is not handled.
Git is checked and handled is not found - COMMIT_ID is `Unknown` is Git
was not found.

BUILD_ID depends on `date` in Linux and `powershell` in Windows, if they
are not found BUILD_ID is set to `Unknown`.